+++ src/main/java/ch/wisv/events/core/model/product/SearchItem.java
@@ -22,15 +22,14 @@ public class SearchItem {
      */
     @Getter
     @Setter
-    public Long data;
+    public Integer data;
 
     /**
      * Default constructor.
-     *
-     * @param title Title of the Product
+     *  @param title Title of the Product
      * @param key   Key of the Product
      */
-    SearchItem(String title, Long key) {
+    SearchItem(String title, Integer key) {
         this.value = title;
         this.data = key;
     }
